+++ NEWS.md
@@ -1,3 +1,5 @@
+# treasury (development version)
+
 # treasury 0.4.0
 
 * Fixed bug in checking for {readxl} installation.